Fixed the files which fail the Scalafmt parser
