+++ math/src/main/scala/breeze/numerics/financial/package.scala
@@ -139,8 +139,8 @@ package object financial {
     val res = roots(reverse(cashflow))
 
     val realRes = DenseVector[Double](
-      for(c:Complex <- res.toArray
-          if ((c.im() == 0) && (0 < c.re())))
+      for (c:Complex <- res.toArray
+          if (c.im() == 0 && c.re() > 0))
       yield c.re()
     )
     val rates = realRes.mapValues(v =>1.0 / v - 1.0)
